requirements.txt
asgiref==3.4.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.12
cur==0.0.1
Django==3.2.13
google-api-core==1.31.5
google-auth==1.35.0
google-cloud-dialogflow==2.12.0
googleapis-common-protos==1.55.0
grpcio==1.45.0
grpcio-status==1.45.0
idna==3.3
Library==0.0.0
mysqlclient==2.1.0
numpy==1.19.5
pandas==1.1.5
Pillow==8.3.2
proto-plus==1.20.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
python-decouple==3.6
djangorestframework==3.13.1
cycler==0.11.0
fonttools==4.27.1
matplotlib==3.3.4
opencv-python==4.5.5.64
packaging==21.3
pyparsing==3.0.9
PyYAML==6.0
scipy==1.5.4
seaborn==0.11.2
torch==1.10.1
torchvision==0.11.2
tqdm==4.64.0
typing_extensions==4.1.1
utils==1.0.1
